services/user_management/app_user_management/requirement.txt
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels-redis==4.2.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==43.0.3
Cython==3.0.11
daphne==4.1.2
Django==4.2.16
django-allauth==65.2.0
django-cleanup==9.0.0
django-cors-headers==4.5.0
django-environ==0.11.2
django-htmx==1.21.0
django-ninja==1.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
msgpack==1.1.0
ninja==1.11.1.1
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyOpenSSL==24.2.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-django==4.9.0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
screeninfo==0.8.1
service-identity==24.2.0
shortuuid==1.0.13
sqlparse==0.5.1
style==1.1.0
tomli==2.1.0
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
update==0.0.1
urllib3==2.2.3
websockets==14.1
whitenoise==6.8.2
zope.interface==7.1.1
